+++ changelog.md
@@ -1,4 +1,4 @@
-## Upcoming
+## 5.5.1.2-alpha
 * Lowered iron armor regenerating absorption
 * `insanesurvivaloverhaul:no_stack_size_changes` has been split into `insanesurvivaloverhaul:no_block_stack_size_changes`, `insanesurvivaloverhaul:no_item_stack_size_changes` and `insanesurvivaloverhaul:no_food_stack_size_changes`
 * Fixed missing copper nugget in creative inventory